requirements.txt
streamlit==1.16.0
nltk==3.7
matplotlib==3.5.0
wordcloud==1.8.2.2
Pillow==8.4.0
numpy==1.21.2
